requirements.txt
Flask
pymongo
requests
opencv-python
mediapipe
tweepy
python-dotenv
